# docs/openapi_v1.json
{
  "openapi": "3.1.0",
  "info": {
    "title": "TraceMind Controller Server API",
    "version": "0.1.0"
  },
  "paths": {
...
    "/api/v1/artifacts": {
      "get": {
        "tags": [
          "artifacts"
        ],
        "summary": "List Artifacts",
        "operationId": "list_artifacts_api_v1_artifacts_get",
        "parameters": [
          {
            "name": "artifact_type",
            "in": "query",
            "required": false,
            "schema": {
              "anyOf": [
                {
                  "type": "string"
                },
                {
                  "type": "null"
                }
              ],
              "title": "Artifact Type"
            }
          },
          {
            "name": "workspace_id",
            "in": "query",
            "required": false,
            "schema": {
              "anyOf": [
                {
                  "type": "string"
                },
                {
                  "type": "null"
                }
              ],
              "description": "Workspace to scope artifacts",
              "title": "Workspace Id"
            },
            "description": "Workspace to scope artifacts"
          }
        ],
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {
                  "type": "array",
                  "items": {
                    "$ref": "#/components/schemas/ArtifactEntryResponse"
                  },
                  "title": "Response List Artifacts Api V1 Artifacts Get"
                }
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      },
      "post": {
        "tags": [
          "artifacts"
        ],
        "summary": "Create Artifact",
        "operationId": "create_artifact_api_v1_artifacts_post",
        "parameters": [
          {
            "name": "workspace_id",
            "in": "query",
            "required": false,
            "schema": {
              "anyOf": [
                {
                  "type": "string"
                },
                {
                  "type": "null"
                }
              ],
              "description": "Workspace to persist the artifact",
              "title": "Workspace Id"
            },
            "description": "Workspace to persist the artifact"
          }
        ],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/ArtifactCreateRequest"
              }
            }
          }
        },
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ArtifactDetailResponse"
                }
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      }
    },
    "/api/v1/artifacts/{artifact_id}": {
      "get": {
        "tags": [
          "artifacts"
        ],
        "summary": "Get Artifact",
        "operationId": "get_artifact_api_v1_artifacts__artifact_id__get",
        "parameters": [
          {
            "name": "artifact_id",
            "in": "path",
            "required": true,
            "schema": {
              "type": "string",
              "title": "Artifact Id"
            }
          },
          {
            "name": "workspace_id",
            "in": "query",
            "required": false,
            "schema": {
              "anyOf": [
                {
                  "type": "string"
                },
                {
                  "type": "null"
                }
              ],
              "description": "Workspace to scope artifacts",
              "title": "Workspace Id"
            },
            "description": "Workspace to scope artifacts"
          }
        ],
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ArtifactDetailResponse"
                }
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      },
      "put": {
        "tags": [
          "artifacts"
        ],
        "summary": "Update Artifact",
        "operationId": "update_artifact_api_v1_artifacts__artifact_id__put",
        "parameters": [
          {
            "name": "artifact_id",
            "in": "path",
            "required": true,
            "schema": {
              "type": "string",
              "title": "Artifact Id"
            }
          },
          {
            "name": "workspace_id",
            "in": "query",
            "required": false,
            "schema": {
              "anyOf": [
                {
                  "type": "string"
                },
                {
                  "type": "null"
                }
              ],
              "description": "Workspace to persist the artifact",
              "title": "Workspace Id"
            },
            "description": "Workspace to persist the artifact"
          }
        ],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/ArtifactCreateRequest"
              }
            }
          }
        },
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ArtifactDetailResponse"
                }
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      }
    },
...
    "/api/controller/artifacts": {
      "get": {
        "tags": [
          "controller"
        ],
        "summary": "List Artifacts",
        "operationId": "list_artifacts_api_controller_artifacts_get",
        "parameters": [
          {
            "name": "intent_id",
            "in": "query",
            "required": false,
            "schema": {
              "anyOf": [
                {
                  "type": "string"
                },
                {
                  "type": "null"
                }
              ],
              "title": "Intent Id"
            }
          },
          {
            "name": "body_hash",
            "in": "query",
            "required": false,
            "schema": {
              "anyOf": [
                {
                  "type": "string"
                },
                {
                  "type": "null"
                }
              ],
              "title": "Body Hash"
            }
          },
          {
            "name": "artifact_type",
            "in": "query",
            "required": false,
            "schema": {
              "anyOf": [
                {
                  "type": "string"
                },
                {
                  "type": "null"
                }
              ],
              "title": "Artifact Type"
            }
          },
          {
            "name": "workspace_id",
            "in": "query",
            "required": false,
            "schema": {
              "anyOf": [
                {
                  "type": "string"
                },
                {
                  "type": "null"
                }
              ],
              "description": "Workspace to query",
              "title": "Workspace Id"
            },
            "description": "Workspace to query"
          }
        ],
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {
                  "type": "array",
                  "items": {
                    "type": "object",
                    "additionalProperties": true
                  },
                  "title": "Response List Artifacts Api Controller Artifacts Get"
                }
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      }
    },
    "/api/controller/artifacts/{artifact_id}": {
      "get": {
        "tags": [
          "controller"
        ],
        "summary": "Get Artifact",
        "operationId": "get_artifact_api_controller_artifacts__artifact_id__get",
        "parameters": [
          {
            "name": "artifact_id",
            "in": "path",
            "required": true,
            "schema": {
              "type": "string",
              "title": "Artifact Id"
            }
          },
          {
            "name": "workspace_id",
            "in": "query",
            "required": false,
            "schema": {
              "anyOf": [
                {
                  "type": "string"
                },
                {
                  "type": "null"
                }
              ],
              "description": "Workspace to read artifacts from",
              "title": "Workspace Id"
            },
            "description": "Workspace to read artifacts from"
          }
        ],
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {
                  "type": "object",
                  "additionalProperties": true,
                  "title": "Response Get Artifact Api Controller Artifacts  Artifact Id  Get"
                }
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      }
    },
...
    "/api/v1/controller/artifacts": {
      "get": {
        "tags": [
          "controller",
          "controller.v1"
        ],
        "summary": "List Artifacts",
        "operationId": "list_artifacts_api_v1_controller_artifacts_get",
        "parameters": [
          {
            "name": "intent_id",
            "in": "query",
            "required": false,
            "schema": {
              "anyOf": [
                {
                  "type": "string"
                },
                {
                  "type": "null"
                }
              ],
              "title": "Intent Id"
            }
          },
          {
            "name": "body_hash",
            "in": "query",
            "required": false,
            "schema": {
              "anyOf": [
                {
                  "type": "string"
                },
                {
                  "type": "null"
                }
              ],
              "title": "Body Hash"
            }
          },
          {
            "name": "artifact_type",
            "in": "query",
            "required": false,
            "schema": {
              "anyOf": [
                {
                  "type": "string"
                },
                {
                  "type": "null"
                }
              ],
              "title": "Artifact Type"
            }
          },
          {
            "name": "workspace_id",
            "in": "query",
            "required": false,
            "schema": {
              "anyOf": [
                {
                  "type": "string"
                },
                {
                  "type": "null"
                }
              ],
              "description": "Workspace to query",
              "title": "Workspace Id"
            },
            "description": "Workspace to query"
          }
        ],
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {
                  "type": "array",
                  "items": {
                    "type": "object",
                    "additionalProperties": true
                  },
                  "title": "Response List Artifacts Api V1 Controller Artifacts Get"
                }
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      }
    },
    "/api/v1/controller/artifacts/{artifact_id}": {
      "get": {
        "tags": [
          "controller",
          "controller.v1"
        ],
        "summary": "Get Artifact",
        "operationId": "get_artifact_api_v1_controller_artifacts__artifact_id__get",
        "parameters": [
          {
            "name": "artifact_id",
            "in": "path",
            "required": true,
            "schema": {
              "type": "string",
              "title": "Artifact Id"
            }
          },
          {
            "name": "workspace_id",
            "in": "query",
            "required": false,
            "schema": {
              "anyOf": [
                {
                  "type": "string"
                },
                {
                  "type": "null"
                }
              ],
              "description": "Workspace to read artifacts from",
              "title": "Workspace Id"
            },
            "description": "Workspace to read artifacts from"
          }
        ],
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {
                  "type": "object",
                  "additionalProperties": true,
                  "title": "Response Get Artifact Api V1 Controller Artifacts  Artifact Id  Get"
                }
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      }
    },
...
  },
  "components": {
    "schemas": {
      "ArtifactCreateRequest": {
        "properties": {
          "artifact_type": {
            "type": "string",
            "title": "Artifact Type"
          },
          "body": {
            "additionalProperties": true,
            "type": "object",
            "title": "Body"
          }
        },
        "type": "object",
        "required": [
          "artifact_type",
          "body"
        ],
        "title": "ArtifactCreateRequest"
      },
      "ArtifactDetailResponse": {
        "properties": {
          "entry": {
            "$ref": "#/components/schemas/ArtifactEntryResponse"
          },
          "document": {
            "$ref": "#/components/schemas/ArtifactDocument"
          }
        },
        "type": "object",
        "required": [
          "entry",
          "document"
        ],
        "title": "ArtifactDetailResponse"
      },
...
      "ArtifactDocument": {
        "properties": {
          "envelope": {
            "additionalProperties": true,
            "type": "object",
            "title": "Envelope"
          },
          "body": {
            "additionalProperties": true,
            "type": "object",
            "title": "Body"
          }
        },
        "type": "object",
        "required": [
          "envelope",
          "body"
        ],
        "title": "ArtifactDocument"
      },
      "ArtifactEntryResponse": {
        "properties": {
          "artifact_id": {
            "type": "string",
            "title": "Artifact Id"
          },
          "artifact_type": {
            "type": "string",
            "title": "Artifact Type"
          },
          "body_hash": {
            "type": "string",
            "title": "Body Hash"
          },
          "path": {
            "type": "string",
            "title": "Path"
          },
          "meta": {
            "additionalProperties": true,
            "type": "object",
            "title": "Meta"
          },
          "version": {
            "type": "string",
            "title": "Version"
          },
          "schema_version": {
            "type": "string",
            "title": "Schema Version"
          },
          "created_at": {
            "type": "string",
            "title": "Created At"
          },
          "status": {
            "type": "string",
            "title": "Status"
          },
          "intent_id": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Intent Id"
          }
        },
        "type": "object",
        "required": [
          "artifact_id",
          "artifact_type",
          "body_hash",
          "path",
          "meta",
          "version",
          "schema_version",
          "created_at",
          "status"
        ],
        "title": "ArtifactEntryResponse"
      },
...
      "HTTPValidationError": {
        "properties": {
          "detail": {
            "items": {
              "$ref": "#/components/schemas/ValidationError"
            },
            "type": "array",
            "title": "Detail"
          }
        },
        "type": "object",
        "title": "HTTPValidationError"
      },
...
      "ValidationError": {
        "properties": {
          "loc": {
            "items": {
              "anyOf": [
                {
                  "type": "string"
                },
                {
                  "type": "integer"
                }
              ]
            },
            "type": "array",
            "title": "Location"
          },
          "msg": {
            "type": "string",
            "title": "Message"
          },
          "type": {
            "type": "string",
            "title": "Error Type"
          }
        },
        "type": "object",
        "required": [
          "loc",
          "msg",
          "type"
        ],
        "title": "ValidationError"
      },
...
    }
  }
}